[REFACTOR] move cube_network to network module
